+++ remoteappmanager/restresources/container.py
@@ -1,10 +1,13 @@
+import socket
 import os
 from datetime import timedelta
+import errno
 
-from tornado import gen
+from tornado import gen, ioloop
+from tornado.httpclient import AsyncHTTPClient, HTTPError
+from tornado.log import app_log
 
 from remoteappmanager.docker.docker_labels import SIMPHONY_NS
-from remoteappmanager.handlers.home_handler import _wait_for_http_server_2xx
 from remoteappmanager.rest import exceptions
 from remoteappmanager.rest.resource import Resource
 from remoteappmanager.docker.container import Container as DockerContainer
@@ -197,7 +200,7 @@ class Container(Resource):
 
         Parameters
         ----------
-        container: Container
+        container: docker.Container
            The container to be connected
         """
         # Note, we use the jupyterhub ORM server, but we don't use it for
@@ -214,3 +217,39 @@ class Container(Resource):
         yield _wait_for_http_server_2xx(
             server_url,
             self.application.file_config.network_timeout)
+
+
+@gen.coroutine
+def _wait_for_http_server_2xx(url, timeout=10):
+    """Wait for an HTTP Server to respond at url and respond with a 2xx code.
+    """
+    loop = ioloop.IOLoop.current()
+    tic = loop.time()
+    client = AsyncHTTPClient()
+
+    while loop.time() - tic < timeout:
+        try:
+            response = yield client.fetch(url, follow_redirects=True)
+        except HTTPError as e:
+            # Skip code 599 because it's expected and we don't want to
+            # pollute the logs.
+            if e.code != 599:
+                app_log.warning("Server at %s responded with: %s", url, e.code)
+        except (OSError, socket.error) as e:
+            if e.errno not in {errno.ECONNABORTED,
+                               errno.ECONNREFUSED,
+                               errno.ECONNRESET}:
+                app_log.warning("Failed to connect to %s (%s)", url, e)
+        except Exception as e:
+            # In case of any unexpected exception, we just log it and keep
+            # trying until eventually we timeout.
+            app_log.warning("Unknown exception occurred connecting to "
+                            "%s (%s)", url, e)
+        else:
+            app_log.info("Server at %s responded with: %s", url, response.code)
+            return
+
+        yield gen.sleep(0.1)
+
+    raise TimeoutError("Server at {} didn't respond in {} seconds".format(
+        url, timeout))
