+++ remoteappmanager/templates/home.html
@@ -33,14 +33,14 @@
           {% endif %}
           {% if info.container == None %}
               <div class="col-sm-2 va">
-                  <button id="{{info.mapping_id}}" name="action" value="start" class="start-button btn btn-primary">Start</button>
+                  <button id="{{info.mapping_id}}" name="action" value="start" class="start-button btn btn-primary"><i class="fa fa-spinner fa-spin" aria-hidden="true" style="display: none"></i> Start</button>
               </div>
           {% else %}
               <div class="col-sm-1 va">
-                  <button name="action" value="view" class="btn btn-success">View</button>
+                  <button id="{{ info.container.url_id }}" name="action" value="view" class="view-button btn btn-success">View</button>
               </div>
               <div class="col-sm-1 va">
-                  <button id="{{ info.container.url_id }}" name="action" value="stop" class="stop-button btn btn-danger">Stop</button>
+                  <button id="{{ info.container.url_id }}" name="action" value="stop" class="stop-button btn btn-danger"><i class="fa fa-spinner fa-spin" aria-hidden="true" style="display: none"></i> Stop</button>
               </div>
           {% endif %}
       </div>
